fixed atoi

+++ ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: rmengelb <rmengelb@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/10/11 13:44:05 by rmengelb      #+#    #+#                 */
-/*   Updated: 2024/10/13 16:10:09 by rein          ########   odam.nl         */
+/*   Updated: 2025/02/22 17:29:04 by rein          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int			i;
+	int			sign;
+	long int	result;
 
 	i = 0;
 	sign = 1;
